services/ocr-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
aiofiles==23.2.1

# OCR
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3

# Shared dependencies
motor==3.3.2
redis==5.0.1
python-json-logger==2.0.7

# HTTP client
httpx==0.26.0
